Moved execution to its own file. Wired up StatementRelevance. Added CQLExecution types
